requirements.txt
discord.py==0.16.12
PyYAML==3.12
numpy==1.14.2
Keras==2.1.6
pythainlp==1.6.0.4
deepcut==0.6.1.0
